#34266 defer applying additional knob settings until after node creation
There are a lot of knob settings that are specific to the particular file_type of the Write node. Previously we were creating the node by passing in all of the settings at once which was generating errors if Nuke tried to set the knob before the file_type had been set. This change creates the node with the file_type setting, and then applies the rest of the settings from the codec_settings hook after.

Closes #6