Improve error handling on the case where Shotgun Create is not installed
DESCRIPTION

When Shotgun Create is not installed and someone try to use this app's command,
it print an error in the console saying that it didn't worked because Shotgun
Create is not installed.

We should not print the callstack in the logs since it looks like something
went wrong even if it's a regular usecase.