Restore the behaviour of the app
JIRA: VMR-2289

DESCRIPTION

In order to not break the previous pipelines, the last changes got isolated
inside of hooks. Configs will have to specify that they want to use the new
implementation of the version submission.

The issue was that the app is still used in a Nuke workflow with the
tk-multi-publish2 app and it not be great to break pipelines and change
the expected behaviour on an updates.

Instead, I implemented the new changes inside of hooks and the person adding
the app to a config can choose which implementation to use