SG-15735 - Support old version of Maya (#28)
DESCRIPTION

The playblast argument list is smaller in older versions of Maya. This pull
request catch the potential IndexError that can happen and gracefully continue.

Extra: It looks like there's have been 5 revisions of doPlayblastArgList and
each revision just appended arguments to the list so it's safe to assume that
if the argument is present, it does what we think it does.